# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Anime World API",
    "description": "API for scraping and retrieving anime information from anime-world.co",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
  "paths": {
    "/search": {
      "get": {
        "summary": "Search for anime",
        "description": "Search for anime by keyword",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "description": "Search keyword",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "title": {
                                "type": "string"
                              },
                              "image": {
                                "type": "string"
                              },
                              "type": {
                                "type": "string"
                              },
                              "languages": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "year": {
                                "type": "string"
                              },
                              "url": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "currentPage": {
                          "type": "integer"
                        },
                        "hasNextPage": {
                          "type": "boolean"
                        },
                        "totalResults": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/info/{contentType}/{animeId}": {
      "get": {
        "summary": "Get anime information",
        "description": "Get detailed information about an anime",
        "parameters": [
          {
            "name": "contentType",
            "in": "path",
            "required": true,
            "description": "Content type (series or movie)",
            "schema": {
              "type": "string",
              "enum": ["series", "movie"]
            }
          },
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "description": "Anime ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        },
                        "image": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "genres": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "languages": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "url": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Anime not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/episodes/{contentType}/{animeId}": {
      "get": {
        "summary": "Get episodes list",
        "description": "Get list of episodes for an anime",
        "parameters": [
          {
            "name": "contentType",
            "in": "path",
            "required": true,
            "description": "Content type (series or movie)",
            "schema": {
              "type": "string",
              "enum": ["series", "movie"]
            }
          },
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "description": "Anime ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "episodes": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "number": {
                                "type": "integer"
                              },
                              "title": {
                                "type": "string"
                              },
                              "languages": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                }
                              },
                              "url": {
                                "type": "string"
                              }
                            }
                          }
                        },
                        "totalEpisodes": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Anime not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/servers/{contentType}/{episodeId}": {
      "get": {
        "summary": "Get episode servers",
        "description": "Get available servers for an episode",
        "parameters": [
          {
            "name": "contentType",
            "in": "path",
            "required": true,
            "description": "Content type (series or movie)",
            "schema": {
              "type": "string",
              "enum": ["series", "movie"]
            }
          },
          {
            "name": "episodeId",
            "in": "path",
            "required": true,
            "description": "Episode ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Unique server identifier"
                          },
                          "name": {
                            "type": "string",
                            "description": "Display name of the server"
                          },
                          "type": {
                            "type": "string",
                            "description": "Type of server (e.g., direct, embed)"
                          }
                        }
                      }
                    }
                  },
                  "example": {
                    "success": true,
                    "data": [
                      {
                        "id": "server1",
                        "name": "Server 1",
                        "type": "direct"
                      },
                      {
                        "id": "server2",
                        "name": "Server 2",
                        "type": "embed"
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Episode not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/sources/{contentType}/{animeId}/{episodeId}": {
      "get": {
        "summary": "Get video sources",
        "description": "Get video sources for an episode",
        "parameters": [
          {
            "name": "contentType",
            "in": "path",
            "required": true,
            "description": "Content type (series or movie)",
            "schema": {
              "type": "string",
              "enum": ["series", "movie"]
            }
          },
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "description": "Anime ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "episodeId",
            "in": "path",
            "required": true,
            "description": "Episode ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "url": {
                            "type": "string",
                            "description": "URL for the video source"
                          },
                          "quality": {
                            "type": "string",
                            "description": "Quality of the video"
                          },
                          "isM3U8": {
                            "type": "boolean",
                            "description": "Indicates if this is an HLS stream"
                          },
                          "isEmbed": {
                            "type": "boolean",
                            "description": "Indicates if this is an embed source"
                          }
                        }
                      }
                    }
                  },
                  "example": {
                    "success": true,
                    "data": [
                      {
                        "url": "https://example.com/video.m3u8",
                        "quality": "auto",
                        "isM3U8": true,
                        "isEmbed": false
                      },
                      {
                        "url": "https://player.example.com/embed?id=123",
                        "quality": "auto",
                        "isM3U8": false,
                        "isEmbed": true
                      }
                    ]
                  }
                }
              }
            }
          },
          "404": {
            "description": "Episode not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/languages/{contentType}/{animeId}": {
      "get": {
        "summary": "Get available languages",
        "description": "Get available languages for an anime",
        "parameters": [
          {
            "name": "contentType",
            "in": "path",
            "required": true,
            "description": "Content type (series or movie)",
            "schema": {
              "type": "string",
              "enum": ["series", "movie"]
            }
          },
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "description": "Anime ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": ["Japanese", "Hindi", "English"]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Anime not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/id/{title}": {
      "get": {
        "summary": "Get anime ID by title",
        "description": "Get anime ID based on the title",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "description": "Title of the anime",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "description": "Anime ID on anime-world.co"
                        }
                      }
                    }
                  },
                  "example": {
                    "success": true,
                    "data": {
                      "id": "demon-slayer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Anime not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          }
        }
      }
    }
  }
}